% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CONTROLS}
\alias{CONTROLS}
\title{The Seven Controls of the Full Conditioning Set}
\format{
A character vector of length 7.
}
\usage{
CONTROLS
}
\description{
The Seven Controls of the Full Conditioning Set
}
\keyword{datasets}
