% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DEFAULT_INTEGRATED}
\alias{DEFAULT_INTEGRATED}
\title{Controls Treated as Integrated of Order One}
\format{
A character vector of length 6.
}
\usage{
DEFAULT_INTEGRATED
}
\description{
On economic grounds, not by pretest.
}
\keyword{datasets}
