% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critvals.R
\docType{data}
\name{DEFAULT_LEVELS}
\alias{DEFAULT_LEVELS}
\title{Significance Levels Reported by Default}
\format{
A numeric vector of length 3.
}
\usage{
DEFAULT_LEVELS
}
\description{
Significance Levels Reported by Default
}
\keyword{datasets}
