% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{rademacher}
\alias{rademacher}
\title{Rademacher Weights}
\usage{
rademacher(n)
}
\arguments{
\item{n}{Length.}
}
\value{
A numeric vector of \code{-1} and \code{1}.
}
\description{
A vector of independent \eqn{\pm 1} draws.
}
\examples{
set.seed(1)
rademacher(8)
}
